shop/pyproject.toml
[tool.poetry]
name = "shop"
version = "0.1.0"
description = "My_Shop"
authors = ["Sergey Solop <solop1992@mail.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "5.1"
djangorestframework = "3.15.2"
gunicorn = "23.0.0"
mysqlclient = "2.2.4"
pillow = "10.4.0"
python-dotenv = "^1.0.1"
setuptools = "72.2.0"
asgiref = "3.8.1"
packaging = "24.1"
sqlparse = "0.5.1"
django-request-logging = "0.7.5"
django-debug-toolbar = "4.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
